update precision arguments

+++ tests/test_strategy.py
@@ -41,14 +41,14 @@ def test_invalid_colossalai(monkeypatch):
 
 @RunIf(colossalai=True)
 def test_colossalai_strategy_with_trainer_by_instance():
-    trainer = Trainer(precision=16, strategy=ColossalAIStrategy())
+    trainer = Trainer(precision="16-mixed", strategy=ColossalAIStrategy())
     assert isinstance(trainer.strategy, ColossalAIStrategy)
     assert isinstance(trainer.strategy.precision_plugin, ColossalAIPrecisionPlugin)
 
 
 @RunIf(colossalai=True)
 def test_colossalai_strategy_with_trainer_by_string():
-    trainer = Trainer(precision=16, strategy="colossalai")
+    trainer = Trainer(precision="16-mixed", strategy="colossalai")
     assert isinstance(trainer.strategy, ColossalAIStrategy)
     assert isinstance(trainer.strategy.precision_plugin, ColossalAIPrecisionPlugin)
 
@@ -80,7 +80,7 @@ def test_gradient_clip_algorithm_error(tmpdir):
         default_root_dir=tmpdir,
         accelerator="gpu",
         devices=1,
-        precision=16,
+        precision="16-mixed",
         strategy="colossalai",
         enable_progress_bar=False,
         enable_model_summary=False,
@@ -99,7 +99,7 @@ def test_colossalai_optimizer(tmpdir):
         default_root_dir=tmpdir,
         accelerator="gpu",
         devices=1,
-        precision=16,
+        precision="16-mixed",
         strategy="colossalai",
         enable_progress_bar=False,
         enable_model_summary=False,
@@ -124,7 +124,7 @@ def test_warn_colossalai_ignored(tmpdir):
         default_root_dir=tmpdir,
         accelerator="gpu",
         devices=1,
-        precision=16,
+        precision="16-mixed",
         strategy="colossalai",
         enable_progress_bar=False,
         enable_model_summary=False,
@@ -218,7 +218,7 @@ def test_multi_gpu_checkpointing(tmpdir):
         max_epochs=1,
         accelerator="gpu",
         devices=2,
-        precision=16,
+        precision="16-mixed",
         strategy="colossalai",
         callbacks=[ck],
         num_sanity_val_steps=0,  # TODO: remove once validation/test before fitting is supported again
@@ -235,7 +235,7 @@ def test_multi_gpu_checkpointing(tmpdir):
 def test_test_without_fit(tmpdir):
     model = ModelParallelClassificationModel()
     dm = ClassifDataModule()
-    trainer = Trainer(default_root_dir=tmpdir, accelerator="gpu", devices=2, precision=16, strategy="colossalai")
+    trainer = Trainer(default_root_dir=tmpdir, accelerator="gpu", devices=2, precision="16-mixed", strategy="colossalai")
 
     # Colossal requires warmup, you can't run validation/test without having fit first
     # This is a temporary limitation
@@ -252,7 +252,7 @@ def test_multi_gpu_model_colossalai_fit_test(tmpdir):
         default_root_dir=tmpdir,
         accelerator="gpu",
         devices=2,
-        precision=16,
+        precision="16-mixed",
         strategy=ColossalAIStrategy(initial_scale=32),
         max_epochs=1,
         num_sanity_val_steps=0,  # TODO: remove once validation/test before fitting is supported again
